Project_2_Expense_tracker/requirement.txt
streamlit == 1.53.1
pandas == 2.3.3
pytest == 9.0.2
fastapi==0.127.0
pydantic == 2.12.5
uvicorn == 0.40.0
mysql-connector-python == 9.5.0
requests == 2.32.4
